Sprint 2 Commit
Created Test database that references the ids of questions in the Question database using dummy data. Edited the create_exam html code so whenever questions are added to a test, it adds them with a unique test id. On the html page, before a test is to created, we added a box that will allow the teacher user to define their own unique test id.